requirements.txt
# Conda installables
arrow
Flask
requests

github3.py
pyOpenSSL

# Flask extensions
Flask-Login
Flask-Mail
Flask-OAuthlib
Flask-PyMongo
Flask-Script
Flask-SSLify
